+++ client/tws_foundation_client/test/definition/addresses_service_test.dart
@@ -1,0 +1,66 @@
+import 'dart:convert';
+
+import 'package:csm_client/csm_client.dart';
+import 'package:test/test.dart';
+import 'package:tws_foundation_client/tws_foundation_client.dart';
+
+void main() {
+  late AddressesServiceBase service;
+  late SetViewOut<Address> viewMock;
+  late SetViewOptions<Address> options;
+
+  setUp(
+    () {
+      List<SetViewOrderOptions> noOrderigns = <SetViewOrderOptions>[];
+      options = SetViewOptions<Address>(false, 10, 1, null, noOrderigns, <SetViewFilterNodeInterface<Address>>[]);
+      viewMock = SetViewOut<Address>(<Address>[], 1, DateTime.now(), 3, 0, 20);
+
+      Client mockClient = MockClient(
+        (Request request) async {
+          JObject jObject = switch (request.url.pathSegments.last) {
+            'view' => SuccessFrame<SetViewOut<Address>>('qTracer', viewMock).encode(),
+            _ => <String, dynamic>{},
+          };
+
+          String object = jsonEncode(jObject);
+          return Response(object, 200);
+        },
+      );
+      service = TWSFoundationSource(
+        true,
+        client: mockClient,
+      ).addresses;
+    },
+  );
+
+  test(
+    'View',
+    () async {
+      MainResolver<SetViewOut<Address>> fact = await service.view(options, '');
+
+      bool passed = false;
+      fact.resolve(
+        decoder: (JObject json) => SetViewOut<Address>.des(json, Address.des),
+        onConnectionFailure: () {},
+        onFailure: (FailureFrame failure, int status) {
+          assert(false, 'server returned a success $status');
+        },
+        onException: (Object exception, StackTrace trace) {
+          assert(false, 'server returned a success');
+        },
+        onSuccess: (SuccessFrame<SetViewOut<Address>> success) {
+          passed = true;
+
+          SetViewOut<Address> fact = success.estela;
+          expect(viewMock.page, fact.page);
+          expect(viewMock.pages, fact.pages);
+          expect(viewMock.records, fact.records);
+          expect(viewMock.creation, fact.creation);
+        },
+      );
+
+      expect(passed, true, reason: 'expected the service returned a success');
+    },
+    timeout: Timeout.factor(5),
+  );
+}